✨ feat: 重构背景脚本，模块化处理Google认证、上下文菜单、消息处理和存储管理
